Example handling of larger profile photos
Once the images in `/images/mpsL` are large enough (120px or wider, 4:3
portrait aspect ratio) we can use the code in this commit to add support
for larger photos on person profile pages, with a nice fallback for
people who only have a small (`/images/mps`) photo.

This commit shouldn’t be merged _until_ those larger photos are
available, otherwise it will result in a big empty space around the
image, because it expects a large image to be available.